+++ DSCPullServerSetup/PublishModulesAndMofsToPullServer.psm1
@@ -174,7 +174,7 @@ function New-ZipFromSource
         else
         {
             Write-Warning -Message "Skipping module overwrite. Module with the name $name already exists."
-            Write-Warning -Message "Please specify -Force to overwrite the module with the local version of the module located in $Source or list names of the modules in ModuleNameList parameter to be packaged from PowerShell module pat instead and remove them from $Source folder"
+            Write-Warning -Message "Please specify -Force to overwrite the module with the local version of the module located in $Source or list names of the modules in ModuleNameList parameter to be packaged from PowerShell module path instead and remove them from $Source folder"
         }
 
         $modules += @("$name")
@@ -298,10 +298,16 @@ function Write-LogEntry
         "$env:SystemDrive\inetpub\wwwroot\PSDSCPullServer\web.config"
 
     .PARAMETER OutputFolderPath
+        Path to the Location where the MOF files should be published.
+        This should be used when the PullServer is a SMB share pull server.
+        (https://docs.microsoft.com/nl-nl/powershell/dsc/pull-server/pullserversmb)
         Defaults to $null
 
     .EXAMPLE
        Get-Module <ModuleName> | Publish-ModuleToPullServer
+
+    .EXAMPLE
+       Get-Module <ModuleName> | Publish-ModuleToPullServer -OutputFolderPath "\\Server01\DscService\Module"
 #>
 function Publish-ModuleToPullServer
 {
@@ -339,15 +345,18 @@ function Publish-ModuleToPullServer
 
     Begin
     {
-        if (-not($OutputFolderPath))
+        if (-not($OutputFolderPath) -or -not (Test-Path -Path $OutputFolderPath))
         {
-            if ( -not(Test-Path $PullServerWebConfig))
+            if (-not(Test-Path $PullServerWebConfig))
             {
-                throw "Web.Config of the pullserver does not exist on the default path $PullServerWebConfig. Please provide the location of your pullserver web configuration using the parameter -PullServerWebConfig or an alternate path where you want to publish the pullserver modules to"
+                throw "Web.Config of the pullserver does not exist on the default path $PullServerWebConfig. Please provide the location of your pullserver web configuration using the parameter -PullServerWebConfig or an alternate path where you want to publish the pullserver modules to. This path should exist."
             }
             else
             {
-                # Pull Server exist figure out the module path of the pullserver and use this value as output folder path.
+                <#
+                    Web.Config of Pull Server found so figure out the module path of the pullserver.
+                    Use this value as output folder path.
+                #>
                 $webConfigXml = [System.Xml.XmlDocument] (Get-Content -Path $PullServerWebConfig)
                 $moduleXElement = $webConfigXml.SelectNodes("//appSettings/add[@key = 'ModulePath']")
                 $OutputFolderPath = $moduleXElement.Value
@@ -370,9 +379,8 @@ function Publish-ModuleToPullServer
     }
     End
     {
-        # Now that all the modules are published generate thier checksum.
+        # Now that all the modules are published generate their checksum.
         New-DscChecksum -Path $OutputFolderPath
-
     }
 }
 
@@ -392,8 +400,17 @@ function Publish-ModuleToPullServer
         Path to the Pull Server web.config file, i.e.
         "$env:SystemDrive\inetpub\wwwroot\PSDSCPullServer\web.config"
 
+    .PARAMETER OutputFolderPath
+        Path to the Location where the MOF files should be published.
+        This should be used when the PullServer is a SMB share pull server.
+        (https://docs.microsoft.com/nl-nl/powershell/dsc/pull-server/pullserversmb)
+        Defaults to $null
+
     .EXAMPLE
         Dir <path>\*.mof | Publish-MOFToPullServer
+
+    .EXAMPLE
+        Dir <path>\*.mof | Publish-MOFToPullServer -OutputFolderPath "\\Server01\DscService\Configuration"
 #>
 function Publish-MOFToPullServer
 {
@@ -410,14 +427,32 @@ function Publish-MOFToPullServer
         [Parameter(Mandatory = $true)]
         [ValidateScript({Test-Path -Path $_ -PathType Leaf})]
         [System.String]
-        $PullServerWebConfig
+        $PullServerWebConfig,
+
+        [Parameter()]
+        [System.String]
+        $OutputFolderPath = $null
     )
 
     Begin
     {
-        $webConfigXml = [System.Xml.XmlDocument] (Get-Content -Path $PullServerWebConfig)
-        $configXElement = $webConfigXml.SelectNodes("//appSettings/add[@key = 'ConfigurationPath']")
-        $OutputFolderPath = $configXElement.Value
+        if (-not($OutputFolderPath) -or -not (Test-Path -Path $OutputFolderPath))
+        {
+            if (-not(Test-Path $PullServerWebConfig))
+            {
+                throw "Web.Config of the pullserver does not exist on the default path $PullServerWebConfig. Please provide the location of your pullserver web configuration using the parameter -PullServerWebConfig or an alternate path where you want to publish the pullserver MOFs to. This Path should exist."
+            }
+            else
+            {
+                <#
+                    Web.Config of Pull Server found so figure out the module path of the pullserver.
+                    Use this value as output folder path.
+                #>
+                $webConfigXml = [System.Xml.XmlDocument] (Get-Content -Path $PullServerWebConfig)
+                $configXElement = $webConfigXml.SelectNodes("//appSettings/add[@key = 'ConfigurationPath']")
+                $OutputFolderPath = $configXElement.Value
+            }
+        }
     }
     Process
     {